Add time only option, update dependencies
This is useful for recurring events (e.g. "I'm free between 8am and 1pm
on Mondays and Fridays"). This is the main distinguisher between this
app and the much more widely used everytimezone, so might as well make
it the default behavior.

One can still use datetime behavior by unticking the `Time only` box.